Possible fix for binoc perma zoom (#4176)
# About the pull request

I have no consistent reproducibility for this but this may fix it? Seems
to occur when an explosion blows up the binocs in the hand of the person
while zoomed.


# Explain why it's good for the game

Bug bad

# Testing Photographs and Procedure
<details>
<summary>Screenshots & Videos</summary>

Put screenshots and videos here with an empty line between the
screenshots and the `<details>` tags.

</details>


# Changelog

:cl: Morrow
fix: Possible fix for binoc perma zoom
/:cl: