ASO/QM Database access (#4230)
# About the pull request
Takes database access from the QM and gives it to the ASO. (Ability to
modify IDs)
<!-- Remove this text and explain what the purpose of your PR is.

Mention if you have tested your changes. If you changed a map, make sure
you used the mapmerge tool.
If this is an Issue Correction, you can type "Fixes Issue #169420" to
link the PR to the corresponding Issue number #169420.

Remember: something that is self-evident to you might not be to others.
Explain your rationale fully, even if you feel it goes without saying.
-->

# Explain why it's good for the game
Oversight that ASO can't modify IDs. Removing from QM as due to place in
command inheritance they shouldn't keep it as those above don't have it.
# Testing Photographs and Procedure
<details>
<summary>Screenshots & Videos</summary>

Put screenshots and videos here with an empty line between the
screenshots and the `<details>` tags.

</details>


# Changelog
:cl:
add: Gave the ASO ability to modify IDs.
del: Removed this from the QM.
/:cl: