Fix add a check when trying to regain control of locked DS to avoid waste of time. (#4209)
# About the pull request
fixes: https://github.com/cmss13-devs/cmss13/issues/4188
add a check before they start the long process of unlock the DS after
queen lock it to avoid people wasting time.

# Explain why it's good for the game
just a little fix or qol or both...
# Testing Photographs and Procedure
<details>
<summary>Screenshots & Videos</summary>

Put screenshots and videos here with an empty line between the
screenshots and the `<details>` tags.


![image](https://github.com/cmss13-devs/cmss13/assets/117036822/424aadaa-536c-4fd1-9ffc-e54c92e2c05d)

</details>


# Changelog
:cl:
fix: add a check before people start the long process of unlock the DS
so that people don't waste time.
/:cl:

---------

Co-authored-by: Julien <jverger.ingx@gmail.com>
Co-authored-by: harryob <me@harryob.live>