attachment obliteration (#4217)
# About the pull request

<!-- Remove this text and explain what the purpose of your PR is.

Mention if you have tested your changes. If you changed a map, make sure
you used the mapmerge tool.
If this is an Issue Correction, you can type "Fixes Issue #169420" to
link the PR to the corresponding Issue number #169420.

Remember: something that is self-evident to you might not be to others.
Explain your rationale fully, even if you feel it goes without saying.
-->
fixes #4194 
Fixes weird behaviour with underbarrel fire extinguisher being able to
be used one handed on some weapons but not on others

All attachable weapons will now require you to hold them with both hands
in order to be used. (( added an exception flag for future attachies
that may need it))

standarized error message you receive when you are not wielding an
attachment.

# Explain why it's good for the game
realism and standarized behaviour for all attachments 

# Testing Photographs and Procedure
<details>
<summary>Screenshots & Videos</summary>

Put screenshots and videos here with an empty line between the
screenshots and the `<details>` tags.

</details>


# Changelog
:cl:
balance: Masterkey and underbarrel extinguisher now must be wielded in
order to be used.
fix: Underbarrel extinguisher no longer bypass wielding restrictions on
flamers.
/:cl: