Ghost scan health nerf round 2 (#4256)
# About the pull request

Sneaky fellas were waiting to be revived and then scanning health before
re-entering the body which bypassed the previous checks. Naughty
naughty.


# Explain why it's good for the game

Not intentional.


# Testing Photographs and Procedure
<details>
<summary>Screenshots & Videos</summary>

Put screenshots and videos here with an empty line between the
screenshots and the `<details>` tags.

</details>


# Changelog

:cl: Morrow
fix: fixed a bypass to allow ghost scan health when it should not be
allowed
/:cl: