actually fixes m4ra custom barrel offset (#4265)
# About the pull request

I'm very confused, one day this is the correct thing that fixes the
offset and the next it breaks it. I literally tested it, but I guess the
old offsets are right? Maybe?

# Explain why it's good for the game


# Testing Photographs and Procedure
<details>
<summary>Screenshots & Videos</summary>

Put screenshots and videos here with an empty line between the
screenshots and the `<details>` tags.

</details>


# Changelog
:cl:
fix: actually fixes m4ra custom barrel offset
/:cl: