Fixes medbay door buttons (#4264)
# About the pull request

Fixes the button IDs on the western medbay doors.

I just want it on the record this is probably the third time I've fixed
these and if I have to do it again I will curse your family for 11
generations.

# Explain why it's good for the game

Bug bad


# Testing Photographs and Procedure
<details>
<summary>Screenshots & Videos</summary>

Put screenshots and videos here with an empty line between the
screenshots and the `<details>` tags.

</details>


# Changelog

:cl: Morrow
fix: Fixed western medbay door buttons
/:cl: