Fixes the wire cutter bug on the plants (#4259)
# About the pull request

Fixes #3367 

In the hydro code for the wire cutter, it forces the update in
process(), so everytime you take a sample there's a chance it grows. So,
by removing the force update it should fix the bug.

# Explain why it's good for the game




# Testing Photographs and Procedure
<details>
<summary>Screenshots & Videos</summary>

Put screenshots and videos here with an empty line between the
screenshots and the `<details>` tags.

</details>


# Changelog



:cl:

fix: Plants should not grow every time it's sampled

/:cl: