Fixes resin areas around LZs (#4210)
# About the pull request
Some areas (such as sorokyne LZ) are weedable when they're not supposed
to be. This fixes those instances.
Primary focus was the sorokyne LZ and dropship console single-tile
areas.
Some areas should be reviewed on if needed where added to Sorokyne.
<!-- Remove this text and explain what the purpose of your PR is.

Mention if you have tested your changes. If you changed a map, make sure
you used the mapmerge tool.
If this is an Issue Correction, you can type "Fixes Issue #169420" to
link the PR to the corresponding Issue number #169420.

Remember: something that is self-evident to you might not be to others.
Explain your rationale fully, even if you feel it goes without saying.
-->

# Explain why it's good for the game
Consistency!
# Testing Photographs and Procedure
<details>
<summary>Screenshots & Videos</summary>

Put screenshots and videos here with an empty line between the
screenshots and the `<details>` tags.

</details>


# Changelog
:cl:
fix: Areas around LZs (particularly on Sorokyne) are now unweedable as
intended.
/:cl: