Lesser drone changes (#4280)
# About the pull request

This PR:

Allows lesser drones to spawn at pylons and core.

Creates a pool of lesser drones at cores and pylons that regenerate over
time. It takes about 10 seconds to get one more when queen is on ovi and
about 120 seconds when queen is off.

Lesser drones and facehuggers no longer count towards the lesser drone
count but lesser drones are still capped at 1/3rd of the hive with a
minimum of 3.

Facehuggers have also been changed to not count lesser drones and
themselves for max calculations.

# Explain why it's good for the game

While fun, the infinite and immediate respawn of lesser drones from hive
core while queen is on ovi is a bit much. This should help with this
problem while also keeping lesser drones as free and horde-like
components of the xeno force.

# Testing Photographs and Procedure
<details>
<summary>Screenshots & Videos</summary>

Put screenshots and videos here with an empty line between the
screenshots and the `<details>` tags.

</details>


# Changelog

:cl: Morrow
add: Allowed lesser drones to spawn at pylons and core.
add: Created a pool of lesser drones at cores and pylons that regenerate
over time. It takes about 10 seconds to get one more when queen is on
ovi and about 120 seconds when queen is off.
balance: Lesser drones and facehuggers no longer count towards the
lesser drone count but lesser drones are still capped at 1/3rd of the
hive with a minimum of 3.
balance: Facehuggers have also been changed to not count lesser drones
and themselves for max calculations.
/:cl: