Add flanking possiblities to the lifeboat by adding two clamp with doors on each exterior side. (#4274)
# About the pull request
adding flanking possibilities to the lifeboat.
north is an abandon section
while south is an hidden bar...
<!-- Remove this text and explain what the purpose of your PR is.

Mention if you have tested your changes. If you changed a map, make sure
you used the mapmerge tool.
If this is an Issue Correction, you can type "Fixes Issue #169420" to
link the PR to the corresponding Issue number #169420.

Remember: something that is self-evident to you might not be to others.
Explain your rationale fully, even if you feel it goes without saying.
-->

# Explain why it's good for the game
lifeboat as been to easy to hold this will offer more opportunity and
make lifeboat more dynamic.
# Testing Photographs and Procedure
<details>
<summary>Screenshots & Videos</summary>


![68747470733a2f2f616666656374656461726330372e626c6f622e636f72652e77696e646f77732e6e65742f6d6462322f696d616765732f3336363937393335372f31363234393337353637372f6d2f6d6170735f6d61705f66696c65735f5553535f416c6d61796572](https://github.com/cmss13-devs/cmss13/assets/117036822/aa0d745e-dc31-4fdc-8f23-db96dee10fbc)


Put screenshots and videos here with an empty line between the
screenshots and the `<details>` tags.

</details>


# Changelog
:cl:
mapadd: adding an an access to the exterior side of each lifeboats.
/:cl:

---------

Co-authored-by: Julien <jverger.ingx@gmail.com>
Co-authored-by: ihatethisengine <treml.treml@yandex.ru>
Co-authored-by: Morrow <darthbane97@gmail.com>