Researchers now spawn with reagent scanner goggles instead of medhuds (#4211)
# About the pull request

see title, this also adds prescription reagent goggles just in case a
researcher spawns with bad eyesight

# Explain why it's good for the game

a little QoL for researchers, saves them the time of needing to grab
reagent scanner goggles from their vendor if they want to do botany or
want to check the chemicals of something without putting it in a
machine.

# Testing Photographs and Procedure
<details>
<summary>Screenshots & Videos</summary>

Put screenshots and videos here with an empty line between the
screenshots and the `<details>` tags.

</details>


# Changelog



:cl: Private Tristan
qol: Researchers now spawn with reagent scanner goggles.
/:cl: