Machete in-hand activation (#4277)
# About the pull request

Trying to clear vines by clicking on them is extremely infuriating due
to their thin hitboxes. This will give marines a much much comfier way
to clear vines, while also making the machete actually be better for
doing it than the easier to store hatchet.
Works by checking the tile directly ahead of the user, and if it finds
any cutable objects it will try to cut them.
The delay on cutting was set in a way so as to reflect the speed at what
a semi-good player would be able to cut at with clicking (accounting for
usual server tick slowdown).

Also there appears to be a ton of invisible vines littered all over LV,
someone should fix that.

# Explain why it's good for the game

QOL, less annoying to do an important job on maps like LV-624.


# Testing Photographs and Procedure
<details>
<summary>Screenshots & Videos</summary>




https://github.com/cmss13-devs/cmss13/assets/15560820/e0ccb094-4f1a-49cd-a8cb-08b4428bfa33




</details>


# Changelog
:cl:
qol: Added in-hand activation for machetes, which will cause one to
start cutting vines and grass on the tile directly in-front.
/:cl:

---------

Co-authored-by: Rigó János <rigojanos@protonmail.com>