Yautja Legacy restoration & Yautja Sprite Fixes (#4033)
# About the pull request
Removes duplicate plasma pistol and thwei sprites.
Renamed one of the legacy mask sprites to what it was supposed to be.
Renamed the health analyzer for preds from Tracker to scanner.
Changed relay beacon to a newer sprite we never used.
Fixes the code for selecting what used to be "elder" armor. Locked it
off to the keys that had that gear.

<!-- Remove this text and explain what the purpose of your PR is.

Mention if you have tested your changes. If you changed a map, make sure
you used the mapmerge tool.
If this is an Issue Correction, you can type "Fixes Issue #169420" to
link the PR to the corresponding Issue number #169420.

Remember: something that is self-evident to you might not be to others.
Explain your rationale fully, even if you feel it goes without saying.
-->

# Explain why it's good for the game
Cleaning out duplicate sprites and old ones is useful I think.
Fixed the legacy gear as we still held the sprites and ~half the code,
so might as well make it work as we do have at least one of the users
still around.
# Testing Photographs and Procedure
<details>
<summary>Screenshots & Videos</summary>

Put screenshots and videos here with an empty line between the
screenshots and the `<details>` tags.
Tested that it works.

![image](https://github.com/cmss13-devs/cmss13/assets/41653574/cc02b39c-5379-4752-adde-43cffb626bc0)


</details>


# Changelog
:cl:
add: Added a preference for using the old legacy yautja armor, only
accessible to the people who had the special gear.
fix: Fixed the old Yautja Legacy armor selection code. Removed the
unique titles from the legacy gear however.
imageadd: Changed the Yautja relay beacon sprite to a newer unused
sprite. Changed iconstate name of health scanner for preds from tracker
to scanner.
imageadd: Moved some of the used sprites out of the unused category in
pred_gear.dmi.
imagedel: Deleted old relay beacon sprite, deleted duplicate plasma
pistol and thwei sprites.
/:cl: