Fixes perma helmet HUDs (#4322)
# About the pull request

Fixes https://github.com/cmss13-devs/cmss13/issues/4317

In the refactor to have all functionality in helmet_visors base it now
requires the active_visor to *not* be the current visor to turn off so
we gotta null active_visor before we turn it off.

# Explain why it's good for the game

Bug bad


# Testing Photographs and Procedure
<details>
<summary>Screenshots & Videos</summary>

Put screenshots and videos here with an empty line between the
screenshots and the `<details>` tags.

</details>


# Changelog

:cl: Morrow
fix: Fixed perma helmet HUDs
/:cl: