Removes debuffs from the folded M39 stock (#4305)
# About the pull request

removes the debuffs from the folded M39 stock, making it equal to a
stockless M39.

# Explain why it's good for the game

currently there's no reason to fold your stock, as it inflicts a heavy
wielded scatter penalty (from 9 scatter when extended to 15 when folded)
with the only benefit being a removal of the wield delay and slightly
better onehanded performance.

a marine shouldn't be penalized for interacting with the folding stock.
hopefully this make the attachment a bit more useful in its secondary
form.

<details>
<summary>stat changes for anyone curious how it looks like
in-game</summary>
  

![image](https://github.com/cmss13-devs/cmss13/assets/17518895/bfd6b1cf-48fa-4fec-b925-8b71b1857112)
(old stock, folded)

![image](https://github.com/cmss13-devs/cmss13/assets/17518895/c7b2aee8-cddd-4e55-b6f1-a005a9ee75ff)
</details>

# Changelog

:cl:
balance: The M39 foldable stock no longer inflicts debuffs when folded.
/:cl: