Adds Helmets + Visor To Synth Vendor | Grey M10 Helmet Addition (#4306)
# About the pull request

Adds several helmet options to the synth vendor (M10, Corpsman and
Technician Helmets), also adds a new 'grey' helmet that can be vended
that is essentially a standard M10 helm but defaulted to the grey/urban
camo. Also adds a riot visor to the vendor.

# Explain why it's good for the game

The helmets are already available to synths across the ship, this just
speeds up the synthetics dressing up time. The grey helmet allows a
synth to match their default synth armour with a helmet no matter the
map.


# Testing Photographs and Procedure
<details>
<summary>Screenshots & Videos</summary>

Grey helmet and visor
![Screenshot 2023-09-02 01 35
54](https://github.com/cmss13-devs/cmss13/assets/6595389/e8b57648-6860-440e-8303-d442fae3fdd8)

Menu
![Screenshot 2023-09-02 01 37
42](https://github.com/cmss13-devs/cmss13/assets/6595389/cca27911-3295-4083-9887-399723f25abf)


</details>


# Changelog
:cl:
add: Synthetic vendors now have the option to vend several helmets and a
riot visor. Including a standard M10 helmet defaulted to non-camo grey.
/:cl:

Co-authored-by: Steelpoint <alexander.henley@hotmai.com>