Royal Marines Commando (#4098)
# About the pull request

This PR adds the Royal Marines Commando under the TWE as a faction and
as an ERT, this PR is part 1 and more is planned to be added the reason
why it isn't being added in this PR is that it isn't fully complete and
these sprites have been waiting a very long time

Royal Marines Roles

Smartgunner: Special smart-gun sprite, gameplaywise only difference is
its use of holo targetting rounds on top of its regular ammo, making it
a menace in a group

Breacher: The breacher combines his high armor with a ballistic shield
and a rifle built to be used in one hand. This guy is hard to bring down

Marksman: The marksman rifle is limited to 20 rounds from 30 but has
increased damage and a mini scope with no slowdown

Rifleman: The grunts of the three world empire are highly maneuverable
their rifle while slower firing and only containing 30 rounds has no
movement slowdown

Teamleader/Officer: the leaders are basically the same as the rifleman
but with better skills and a VP78

The Royal Marines are equipped for every possibility in the field, they
are all trained medics and carry the equipment to boot. Aside from the
breacher, they are all quick on their feet, and their weapons while
across the board slower than their marine equivalents pack the punch
needed to get the job done.

Fluff, the royal marines in this area are based out of the HMS Patna
working alongside the HMS Thunderchild to assist the United Americas in
protecting the outer colonies from increasing CLF activity

# Explain why it's good for the game

Gives the TWE some representation, adds some cool new faction stuff, an
ERT, and equipment

# Testing Photographs and Procedure
<details>
<summary>Screenshots & Videos</summary>

Put screenshots and videos here with an empty line between the
screenshots and the `<details>` tags.

</details>


# Changelog
:cl: Triiodine, Frans, Warfan, SpartanBobby, Twomper, Roboticpotato,
Tophat Penguin,forest2001
add: Added Royal Marines commando ERT and faction
add: Updates Riot shield sprite, adds metal shield as well (sprites from
trio)
add: TWE ERT station
add: F903 Rifle, Royal marines commando armour sets, new smartgun with
holotargetting rounds for royal marines commando
add: Calico shotgun, pistol and SMG sprites to the codebase made by
Triodine
add: royal marines commando HEDP and HIDP grenades
/:cl:

---------

Co-authored-by: Morrow <darthbane97@gmail.com>
Co-authored-by: harryob <me@harryob.live>