On-screen messages sounds (#4301)
# About the pull request

On-screen messages now appear with sound. Radio static for marines,
distant roar for xenos.

# Explain why it's good for the game

Helps to draw attention to the message. Weather broadcast has sounds,
but the rest of the messages for some reason do not.


# Testing Photographs and Procedure
<details>
<summary>Screenshots & Videos</summary>

Put screenshots and videos here with an empty line between the
screenshots and the `<details>` tags.

</details>


# Changelog
:cl: ihatethisengine
add: On-screen messages now appear with sound. Radio static for marines,
distant roar for xenos.
/:cl:

---------

Co-authored-by: ihatethisengine <treml.treml@yandex.ru>
Co-authored-by: Drathek <76988376+Drulikar@users.noreply.github.com>