End game timing changes (#4353)
# About the pull request

This PR:

Lowers the xeno endgame start to 55 minutes into the round.

Lowers the nuke endgame start to 115 minutes into the round.

Lowers the nuke cost to 5 points.

# Explain why it's good for the game

Xenos need a bit of a boost come midgame.

Nuke should be easier to acquire once it is time to consider it and it
should become available on the hour rather than 5 minutes offset due to
roundstart time.

# Testing Photographs and Procedure
<details>
<summary>Screenshots & Videos</summary>

Put screenshots and videos here with an empty line between the
screenshots and the `<details>` tags.

</details>


# Changelog

:cl: Morrow
balance: Lowered the xeno endgame start to 55 minutes into the round
balance: Lowered the nuke endgame start to 115 minutes into the round
balance: Lowered the nuke cost to 5 points
/:cl: