Makes gau able to hit resting mobs (#4304)
# About the pull request

GAU should be able to hit resting mobs, that is just a bug. Copy of a
part of Tisx GAU buff that makes GAU able to hit mobs.

# Explain why it's good for the game

it might make GAU a bit less of a joke. if it is too strong they way to
balance it migth be removal of the explosion and spread decreas by one.



# Changelog



:cl:

balance: GAU can hit mobs that are on ground now

/:cl: