Project ARES: Adds more APOLLO consoles (#4208)
# About the pull request
Adds some APOLLO consoles to medbay (Reception desk, Surgical storeroom
& Research Hydroponics)
Moves the console outside ARES reception to be within the area, propped
on the reception desk.
<!-- Remove this text and explain what the purpose of your PR is.

Mention if you have tested your changes. If you changed a map, make sure
you used the mapmerge tool.
If this is an Issue Correction, you can type "Fixes Issue #169420" to
link the PR to the corresponding Issue number #169420.

Remember: something that is self-evident to you might not be to others.
Explain your rationale fully, even if you feel it goes without saying.
-->

# Explain why it's good for the game
Medical tends to get a bit bloody, and they're the only department
without one (that needs one frequently)
The console outside ARES reception can't be used to return access
tickets because it's outside the area they're automatically removed.
This fixes that.
# Testing Photographs and Procedure
<details>
<summary>Screenshots & Videos</summary>

Put screenshots and videos here with an empty line between the
screenshots and the `<details>` tags.

</details>


# Changelog
:cl:
maptweak: Adds a couple APOLLO consoles to upper and lower medical.
Moves the one outside ARES reception slightly.
/:cl: