Cryo spec one time tech purchase (#4354)
# About the pull request

This PR puts cryo specialist in its own tech.

It can be purchased once for eight points for a single specialist. Point
cost may be changed.

# Explain why it's good for the game

Moving this to its own category gives commanders a choice between a few
men and a specialist while also stopping tons of spec stacking.

# Testing Photographs and Procedure
<details>
<summary>Screenshots & Videos</summary>

Put screenshots and videos here with an empty line between the
screenshots and the `<details>` tags.

</details>


# Changelog

:cl: Morrow
add: Added cryo specialist tech that can be purchased once for eight
points
/:cl: