Fixes a minor error in acid runner's description (#4414)
# About the pull request
"Corrosive Acid equivalent to that of a Praetorian" -> " Corrosive Acid
equivalent to that of a boiler"

# Explain why it's good for the game

uh.... description being accurate.... good???? maybe??? 
# Testing Photographs and Procedure
<details>
<summary>Screenshots & Videos</summary>

Put screenshots and videos here with an empty line between the
screenshots and the `<details>` tags.

</details>


# Changelog
:cl: Private Tristan
spellcheck: acid runners are now properly informed of their corrosive
acid's strength.
/:cl: