New facehugger strain: Watcher (#3726)
Super easy to understand:

The watcher loses the ability to hide in exchange for letting them see
further.

I don't have a sprite for this at the moment and probably won't for this
PR so I have made them stand out by making them more orange because I
think marines being able to see if a hugger is a Watcher might be
important.

![image](https://github.com/cmss13-devs/cmss13/assets/24533979/583b12f7-fef3-4da7-a989-14661cf98f5f)

Currently they see 2 extra tiles like the runner but in testmerging I
think it may change to 3 tiles (view range 10 instead of view range 9)
unless I hit the perfect sweet spot already. **EDIT: This number is now
3**

**the way I see it watcher is designed for 2 main playstyles:**

1: Backlining to try to get someone not paying attention or busy (right
behind marine frontline like medics doing medic stuff or someone
resupplying/playing with inventory) or deep in the backline to try to
find someone vulnerable.

2: Hanging around the xeno side frontline with increased situational
awareness to be exactly where you need to be for xeno assisted caps. For
example as a watcher during the TM I have an easier time watching a
warrior fling someone and capitalizing a hug off a fling because I don't
have to go out until the fling happens etc


update:
------
The activation of the view range now works similar to the runner; it is
toggle-able

![dreamseeker_tyE3nVA9V0](https://github.com/cmss13-devs/cmss13/assets/24533979/b232b076-8f0b-4a83-abbc-fe4c84a72d21)

Fixed exploit that allowed watchers to hide despite not having the
ability.

Update 2:
--------
The watcher now does not take weeds over time. 
Essentially implemented this:


![Discord_9s19Vo2olV](https://github.com/cmss13-devs/cmss13/assets/24533979/0f2d3121-2c3d-4751-b8e9-263df9a9a697)

**Why:** 
The damage over time is meant so that the hugger cannot hide under some
unweeded item the whole round to hug some random unsuspecting marine.
The watcher cannot hide and is meant to be a backlining caste.
This allows the watcher to stalk people outside of weeds and wait for
the perfect opportunity to hug when their target is distracted or busy.

Update 3:
--------
Added my codersprite to the watcher.
These were created by me so hopefully a spriter or my spriter replaces
these with a better one if they suck.


![dreamseeker_O6I1iXuo21](https://github.com/cmss13-devs/cmss13/assets/24533979/55ddf0cb-a23a-448a-9afa-d8d987fa3d03)




:cl: Hopek
add: New facehugger strain: Watcher. The Watcher loses their ability to
hide in exchange to see further and the ability to no longer take damage
outside of weeds.
/:cl:

---------

Co-authored-by: ihatethisengine <115417687+ihatethisengine@users.noreply.github.com>