IOs get their own squads (#4375)
# About the pull request

IOs get their own squad, and they get assigned roundstart

# Explain why it's good for the game

It promotes teamwork for IOs, the lack of which is the
https://github.com/cmss13-devs/cmss13/pull/1 cause of IO death

In addition there's already an overwatch console in the CIC office.

You could already do this via Echo, but that's not apparant to new
players

# Testing Photographs and Procedure
<details>
<summary>Screenshots & Videos</summary>

Put screenshots and videos here with an empty line between the
screenshots and the `<details>` tags.

</details>


# Changelog
:cl: TheGamerdk, ihatethisengine
add: IOs start in their own squad now
/:cl:

---------

Co-authored-by: TheGamerdk <5618080+TheGamerdk@users.noreply.github.com>
Co-authored-by: ihatethisengine <treml.treml@yandex.ru>