Science Annex Yard Readdition (#4379)
# About the pull request

This PR aims to readd yard to Science Annex. In February last year, we
removed yard due to having a negative effect on Science Annex's gameplay
flow. However, the complete removal of it has shown over time that it
also has its own major issues. This readdition should address the flow
issues that the map suffers.

Will probably require test merging to gauge its effect, another possible
solution is only allowing it to serve as a connector to the North and
South.

# Explain why it's good for the game

Science Annex currently suffers badly from there being no center of the
map, meaning that you have to travel all the way around the map to reach
the other side. This should aid in the flow of the map, without allowing
yard to turn into the meta afk and hold spot while ignoring the rest of
the map. You can use yard as a hold, but you will need to use the rest
of the map to push.


# Testing Photographs and Procedure
<details>
<summary>Screenshots & Videos</summary>


![image](https://github.com/cmss13-devs/cmss13/assets/133173804/1009ca47-3119-4076-a221-df84122784d5)


</details>


# Changelog

:cl:
mapadd: Readds yard to Science Annex
maptweak: Redid the yard loot to be more in line with current surv loot
standards.
/:cl:

---------

Co-authored-by: Morrow <darthbane97@gmail.com>